Add a master controller light
