+++ tests/smart_home.py
@@ -20,3 +20,4 @@ dmx.web_control()
 # Close the controller once we're done
 dmx.sleep_till_enter()
 dmx.close()
+smart.stop()
